requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
commonmark==0.9.1
idna==3.4
ping3==4.0.3
python-dotenv==0.14.0
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.6.0
urllib3==1.26.13